Fix issues with volume plot. Revert previous change that added avtVolumeResampleFilter to ApplyOperators and put it back to using avtResampleFilter in ApplyRenderingTransform. With that change, removed test and Exception for Rectilinear grids with implied transform, as they are handled just fine by the Resampling. Added more filters to ReleaseData, resolves issue with Composite rendering messing up other types of rendering.
