+++ src/plots/VolumeVTK9/Volume.xml
@@ -19,7 +19,6 @@
       avtVisItVTKRenderer.C
       avtVisItVTKRenderFilter.C
       avtICETCompositor.C
-      avtVolumeResampleFilter.C
     </Files>
     <LIBS components="V">
       VTK::RenderingVolume
